fix:unauthorized error

+++ app/api/admin/enquiry/route.ts
@@ -1,10 +1,15 @@
 import connectDB from "@/lib/mongodb";
 import { NextRequest, NextResponse } from "next/server";
 import Enquiry from "@/models/Enquiry";
+import { verifyAdmin } from "@/lib/verifyAdmin";
 
 export async function POST(request: NextRequest) {
     try {
         await connectDB();
+        const isAdmin = await verifyAdmin(request);
+        if (!isAdmin) {
+            return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+        }
         const body = await request.json();
         const enquiry = await Enquiry.create(body);
         if(!enquiry){
@@ -35,6 +40,10 @@ export async function GET() {
 export async function DELETE(request: NextRequest) {
     try {
         await connectDB();
+        const isAdmin = await verifyAdmin(request);
+        if (!isAdmin) {
+            return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
+        }
         const id = request.nextUrl.searchParams.get("id");
         const enquiry = await Enquiry.findByIdAndDelete(id);
         if (!enquiry) {
